Improve the nucleotide profile-based filtering
Tests from the test_hypothesis_profile branch confirm the validity of
the nucleotide profile-based filtering, and indicate that it is
possible to use the length difference between the two considered
amplicons to improve the filtering threshold (just subtract the
length difference).

The improvement is small. The number of comparisons avoided decreases
by 0.6%. Nevertheless, I keep it since it adds no computational
overhead and can easily be reverted.