Use Needleman-Wunsch instead of Levenshtein
Previous versions of swarm.py were based on a Levenshtein algorithm
(edition distance) implemented by Damien Miller as a python module
written in C (https://code.google.com/p/py-editdist/). To improve the
accuracy of pairwise amplicon comparisons, it uses a pure python
Needleman-Wunsch algorithm to get the optimal alignment (minimizing
the number of penalties). It then back-traces the alignment to count
the number of mismatches.

The model used is adapted from a model with +5 for a match, -4 for a
mismatch, -12 for a gap opening and -4 for a gap extension. That model
is transformed into a pure penalty model: 0 for a match, 9 for a
mismatch or a gap extension and 21 for a gap opening. That can be
simplified by a factor 3: 0 for a match, 3 for a mismatch or a gap
extension, 7 for a gap opening. That transformation improves the
performances of the C++ swarm implementation (score stored on a single
byte of memory).

This version is now almost similar to swarm (1.0.2), but much slower
(and also much slower than previous versions of swarm.py). That
sluggishness is not really a problem since this python version is
considered as a sand box designed to easily test ideas. The difference
between the python and the C++ versions of swarm comes from the use of
profile-based filtering in the python version. That optimization will
be implemented in swarm.cpp soon.