Use the third-party python module bitarray to store and compare kmer-vectors
We've gained some speed with the re-write of swarm. The pure
Needleman-Wunsch version was running in 130 s. Introducing the
filtering with 5-mers shave off 15 s. Using bitarray instead of a pure
python Boolean array shave off 30 s. Now, profiling shows that 99% of
the computation time is spent in the Needleman-Wunsch function.